@@ -1,6 +1,5 @@
 import React, { Component } from "react";
-import {StyleSheet, Text, View} from "react-native";
-import { Button } from "react-native-elements";
+import { StyleSheet, Text, View, Button } from "react-native";
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-native';
 
